python-nlp/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
spacy==3.7.2
scikit-learn==1.4.0
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf==4.0.1
numpy==1.26.3
python-dotenv==1.0.0
